+++ NEWS.md
@@ -24,7 +24,7 @@
 
 * `plot.pwsd()` now explicitly includes an option to customize title with `main = .`
 
-* `hhj()` now includes a warning message if the supplied iteration limit `n_iter` is reached.
+* `hhj()` now includes a warning message if the supplied iteration limit `n_iter` is reached but still outputs an object of class 'hhj.'
 
 
 # blocklength 0.1.3